feat: add OpenRouter official pricing integration and data structures
