+++ frontend/features/admin/api/billing.ts
@@ -15,6 +15,7 @@ import type {
   AdminModelPricingDTO,
   AdminModelPricingData,
   AdminModelPricingPage,
+  AdminOfficialPricingCatalogData,
   CreateAdminRedemptionCodeRequest,
   UpdateAdminRedemptionCodeRequest,
   UpdateAdminBillingConfigRequest,
@@ -180,6 +181,22 @@ export async function upsertAdminModelPricing(
   return authedRequest<AdminModelPricingData>(
     "/api/v1/admin/billing/model-prices",
     { method: "PUT", accessToken, body: payload },
+    true,
+  );
+}
+
+export async function getAdminOpenRouterOfficialPricing(
+  accessToken: string,
+  options: { refresh?: boolean } = {},
+): Promise<AdminOfficialPricingCatalogData> {
+  const params = new URLSearchParams();
+  if (options.refresh) {
+    params.set("refresh", "true");
+  }
+  const suffix = params.size > 0 ? `?${params.toString()}` : "";
+  return authedRequest<AdminOfficialPricingCatalogData>(
+    `/api/v1/admin/billing/official-pricing/openrouter${suffix}`,
+    { accessToken },
     true,
   );
 }
