+++ frontend/features/admin/components/sections/billing/billing-dialogs.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import * as React from "react";
-import { Plus, Trash2 } from "lucide-react";
+import { Plus, Sparkles, Trash2 } from "lucide-react";
 import { motion } from "motion/react";
 import { useTranslations } from "next-intl";
 
@@ -253,6 +253,7 @@ type PricingBillingDialogProps = {
   onAddTier: () => void;
   onRemoveTier: (index: number) => void;
   onUpdateTier: (index: number, patch: Partial<TieredPricingTierForm>) => void;
+  onOpenOfficialPricing: () => void;
 };
 
 export function PricingBillingDialog({
@@ -266,26 +267,24 @@ export function PricingBillingDialog({
   onAddTier,
   onRemoveTier,
   onUpdateTier,
+  onOpenOfficialPricing,
 }: PricingBillingDialogProps) {
   const t = useTranslations("adminBilling");
   const tActions = useTranslations("common.actions");
   const [editorMode, setEditorMode] = React.useState<"form" | "json">("form");
   const [jsonDraft, setJSONDraft] = React.useState("");
   const [jsonError, setJSONError] = React.useState("");
-  const lastSyncedJSONRef = React.useRef("");
 
   React.useEffect(() => {
     if (!open || !form) {
       setEditorMode("form");
       setJSONDraft("");
       setJSONError("");
-      lastSyncedJSONRef.current = "";
       return;
     }
     const nextJSON = pricingFormToJSON(form);
     if (editorMode !== "json") {
       setJSONDraft(nextJSON);
-      lastSyncedJSONRef.current = nextJSON;
       setJSONError("");
     }
   }, [editorMode, form, open]);
@@ -303,7 +302,6 @@ export function PricingBillingDialog({
         tiered: t("modelPricing.jsonErrors.tiered"),
       });
       setJSONError("");
-      lastSyncedJSONRef.current = value;
       setForm(nextForm);
     } catch (error) {
       setJSONError(error instanceof Error ? error.message : t("modelPricing.jsonErrors.invalid"));
@@ -320,7 +318,6 @@ export function PricingBillingDialog({
     setForm(nextForm);
     setJSONDraft(nextJSON);
     setJSONError("");
-    lastSyncedJSONRef.current = nextJSON;
   }, [form, setForm]);
 
   return (
@@ -345,8 +342,21 @@ export function PricingBillingDialog({
               <>
               <div className="space-y-1">
                 <p className="text-xs text-muted-foreground">{t("modelPricing.platformModel")}</p>
-                <div className="flex h-8 items-center rounded-md border border-input/40 bg-muted/30 px-3 text-xs">
-                  <span className="truncate">{form.platformModelName}</span>
+                <div className="flex items-center gap-2">
+                  <Input value={form.platformModelName} className="min-w-0 flex-1 cursor-default text-foreground placeholder:text-muted-foreground" readOnly />
+                  <Button
+                    type="button"
+                    variant="secondary"
+                    size="sm"
+                    className="h-8 shrink-0 px-2.5 text-xs shadow-none"
+                    disabled={saving}
+                    onClick={onOpenOfficialPricing}
+                    aria-label={t("modelPricing.officialPricing")}
+                    title={t("modelPricing.officialPricing")}
+                  >
+                    <Sparkles className="size-3.5 stroke-1" />
+                    {t("modelPricing.officialPricing")}
+                  </Button>
                 </div>
               </div>
