+++ frontend/i18n/messages/en-US/admin-billing.json
@@ -171,6 +171,8 @@
     "basePrice": "Base price",
     "updatedAt": "Updated at",
     "empty": "No model pricing settings",
+    "modelInfo": "Model info",
+    "similarity": "Similarity",
     "dialogTitle": "Model base price",
     "dialogDescription": "Configure the platform model billing mode and base unit prices.",
     "editorMode": "Edit mode",
@@ -184,6 +186,21 @@
       "mode": "pricingMode must be token, call, duration, or tiered.",
       "tiered": "Tiered pricing requires a non-empty tieredPricing.tiers array."
     },
+    "officialPricing": "Quick setup",
+    "officialPricingSingleTitle": "Quick setup",
+    "officialPricingSingleDescription": "Quickly configure model pricing from OpenRouter's model list.",
+    "officialPricingSearchPlaceholder": "Search model pricing",
+    "officialPricingMultiplier": "Import multiplier",
+    "officialPricingSync": "Sync",
+    "officialPricingEmpty": "No matching price candidates found. Try searching manually.",
+    "officialPricingImportTitle": "Import model pricing",
+    "officialPricingImportDescription": "Set an import multiplier before filling the current pricing form.",
+    "officialPricingImport": "Import",
+    "priceInput": "Input",
+    "priceOutput": "Output",
+    "priceCacheWrite": "Cache write",
+    "priceCacheRead": "Cache read",
+    "officialPricingRemoteError": "Live fetch failed. Try syncing manually.",
     "freeModel": "Free model",
     "inputPerM": "Input / 1M",
     "outputPerM": "Output / 1M",
@@ -260,6 +277,10 @@
     "nativeToolBillingSaveFailed": "Failed to save native tool billing",
     "pricingSaved": "Model pricing saved",
     "pricingSaveFailed": "Could not save model pricing",
+    "officialPricingApplied": "Price filled",
+    "officialPricingRemoteLoaded": "Loaded {count} OpenRouter model prices",
+    "officialPricingRemoteFailed": "Failed to fetch OpenRouter pricing",
+    "officialPricingRemoteEmpty": "OpenRouter model list is empty",
     "exportEmpty": "No model pricing to export",
     "exported": "Exported {count} model pricing entries",
     "importUnknownModels": "Import failed: unknown platform models",
