+++ frontend/i18n/messages/zh-CN/admin-billing.json
@@ -171,6 +171,8 @@
     "basePrice": "基础单价",
     "updatedAt": "更新时间",
     "empty": "暂无模型定价配置",
+    "modelInfo": "模型信息",
+    "similarity": "相似度",
     "dialogTitle": "模型基础价格",
     "dialogDescription": "配置平台模型的计费方式与基础单价。",
     "editorMode": "编辑方式",
@@ -184,6 +186,21 @@
       "mode": "pricingMode 必须是 token、call、duration 或 tiered。",
       "tiered": "阶梯计费需要包含非空 tieredPricing.tiers。"
     },
+    "officialPricing": "快速配置",
+    "officialPricingSingleTitle": "快速配置",
+    "officialPricingSingleDescription": "从 OpenRouter 官方模型列表快速配置模型定价",
+    "officialPricingSearchPlaceholder": "搜索模型定价",
+    "officialPricingMultiplier": "导入倍率",
+    "officialPricingSync": "同步",
+    "officialPricingEmpty": "未发现匹配的价格候选，请尝试手动搜索",
+    "officialPricingImportTitle": "导入模型定价",
+    "officialPricingImportDescription": "设置导入倍率后覆盖当前模型定价表单。",
+    "officialPricingImport": "导入",
+    "priceInput": "输入",
+    "priceOutput": "输出",
+    "priceCacheWrite": "缓存写",
+    "priceCacheRead": "缓存读",
+    "officialPricingRemoteError": "实时获取失败，请尝试手动同步。",
     "freeModel": "免费模型",
     "inputPerM": "输入 / 1M",
     "outputPerM": "输出 / 1M",
@@ -260,6 +277,10 @@
     "nativeToolBillingSaveFailed": "原生工具计费保存失败",
     "pricingSaved": "模型基础价格已保存",
     "pricingSaveFailed": "保存模型基础价格失败",
+    "officialPricingApplied": "已填入价格",
+    "officialPricingRemoteLoaded": "已获取 {count} 个 OpenRouter 模型价格",
+    "officialPricingRemoteFailed": "获取 OpenRouter 价格失败",
+    "officialPricingRemoteEmpty": "OpenRouter 模型列表为空",
     "exportEmpty": "暂无可导出的模型定价",
     "exported": "已导出 {count} 个模型定价",
     "importUnknownModels": "导入失败：存在未知平台模型",
